Variational methods merged
