--- src/methods/bayes/base/output.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-import torch
-
-
-@dataclass
-class BaseOutputModel:
-    output: torch.Tensor
